Abort install/uninstall scripts if 'leo-editor' is in sys.path

+++ leo/scripts/uninstall_leo.py
@@ -5,21 +5,30 @@
 """
 uninstall_leo.py: Run `pip uninstall leo`.
 
+*Note*: The leo-editor folder must *not* be in sys.path!
+
 Info item #3837 describes all distribution-related scripts.
 https://github.com/leo-editor/leo-editor/issues/3837
 """
-
 import os
+import sys
 import subprocess
 
-print(os.path.basename(__file__))
+file_name = os.path.basename(__file__)
 
-# Do *not* install from leo-editor!
-home_dir = os.path.expanduser("~")
-os.chdir(home_dir)
+if any('leo-editor' in z for z in sys.path):
+    print(f"{file_name}: remove leo-editor from sys.path!")
+    print('Hint: do *not* run this script from the leo-editor directory!')
+else:
+    print(os.path.basename(__file__))
 
-# Uninstall.
-command = 'python -m pip uninstall leo'
-print(command)
-subprocess.Popen(command, shell=True).communicate()
+    # Do *not* install from leo-editor!
+    home_dir = os.path.expanduser("~")
+    os.chdir(home_dir)
+
+    # Uninstall.
+    # --yes: Don’t ask for confirmation of uninstall deletions.
+    command = 'python -m pip uninstall leo'
+    print(command)
+    subprocess.Popen(command, shell=True).communicate()
 #@-leo
